Add subtractions
